Clist modern: Minor cleanup
git-svn-id: http://svn.miranda-ng.org/main/trunk@12632 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Clist_modern/src/hdr/modern_clist.h
@@ -85,7 +85,7 @@ struct ClcCacheEntry : public ClcCacheEntryBase
 	int      NotOnList;
 	int      IdleTS;
 	void*    ClcContact;
-	BYTE     IsExpanded;
+	int      IsExpanded;
 
 	TCHAR*   szSecondLineText;
 	CSmileyString ssSecondLine;
